requirements.txt
# backend
Flask==3.1.2
requests==2.32.0
Flask-Limiter==2.10.4
python-dotenv==1.0.0

# frontend
gradio==3.39.0